Use .js extension instead of .ts
This was such a headache to figure out. I had to change to tsx from tsc-node
because it would throw errors when building with .ts extensions
and then throw errors when running start:dev with .js extensions
and then throwing more errors when using no extensions

Fuck this ecosystem. Or am I just stupid?